rearranges an renames scripts
