+++ dags/dag_content/nextflow_tower_functions.py
@@ -13,20 +13,8 @@ def create_and_open_tower_workspace(platform: str, workspace_id: str) -> TowerUt
         tower_utils (TowerUtils): Authenticated Tower Utils class
     
     """
-    config_dict = {
-        #"sage": {
-        # "token": Variable.get("TOWER_ACCESS_TOKEN"),
-        # } placeholder for prod platform, will need refactor with migration
-        "sage-dev": {
-            "token": Variable.get("TOWER_ACCESS_TOKEN"),
-        }
-    }
-
-    if platform not in config_dict.keys():
-        raise KeyError(f"Platform '{platform}' is not currently supported. Please select from {str(list(config_dict.keys()))}")
-    
     client_args = TowerUtils.bundle_client_args(
-        auth_token=config_dict[platform]["token"], platform=platform
+        auth_token=Variable.get("TOWER_ACCESS_TOKEN"), platform=platform
     )
     tower_utils = TowerUtils(client_args)
     tower_utils.open_workspace(workspace_id=workspace_id)
