+++ dags/tower_metrics_dag.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 from airflow.decorators import dag
 
-from dag_content.tower_metrics_content import *
+from dag_content.tower_metrics_functions import *
 
 
 @dag(
